Start moving stuff to a Host namespace
