+++ XAMLTest/Internal/App.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using XamlTest.Host;
 
 namespace XamlTest.Internal
 {
