+++ XAMLTest/Internal/ManagedApp.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
+using XamlTest.Host;
 
 namespace XamlTest.Internal
 {
