+++ XAMLTest/Internal/Window.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
+using XamlTest.Host;
 
 namespace XamlTest.Internal
 {
