+++ XAMLTest/XAMLTest.csproj
@@ -20,7 +20,7 @@
     <PackageReference Include="PInvoke.User32" Version="0.7.104" />
     <PackageReference Include="System.Drawing.Common" Version="5.0.2" />
     <ProjectReference Include="..\XAMLTest.Generator\XAMLTest.Generator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
-    <Protobuf Include="VTspec.proto" GrpcServices="Both" />
+    <Protobuf Include="Host\VTspec.proto" GrpcServices="Both" />
   </ItemGroup>
 
   <ItemGroup>
